added command 'fabric-cli configure asset add'
