+++ samples/fabric/clis/go/fabric-cli/.env.template
@@ -1,0 +1,6 @@
+DEFAULT_CHANNEL=mychannel
+DEFAULT_CHAINCODE=interop
+MEMBER_CREDENTIAL_FOLDER=PATH-TO-CREDENTIAL-FOLDER
+LOCAL=true
+DEFAULT_APPLICATION_CHAINCODE=simplestate
+CONFIG_PATH=PATH-TO-CONFIG-FILE
