+++ samples/fabric/clis/go/fabric-cli/data/assets.json
@@ -1,0 +1,18 @@
+{
+  "a03": {
+    "assetType": "bond01",
+    "id": "a03",
+    "owner": "User1@org1.network1.com",
+    "issuer": "treasury",
+    "facevalue": "500",
+    "maturitydate": "01 Jan 22 00:00 MST"
+  },
+  "a04": {
+    "assetType": "bond01",
+    "id": "a04",
+    "owner": "Admin@org1.network1.com",
+    "issuer": "treasury",
+    "facevalue": "700",
+    "maturitydate": "01 Jul 22 00:00 MST"
+  }
+}
